split up all of the plotting
